Refactor the VulnerabilityAnalysis for option product context #98
