+++ product_portfolio/models.py
@@ -45,6 +45,8 @@ from dje.validators import validate_version
 from vulnerabilities.fetch import fetch_for_queryset
 from vulnerabilities.models import Vulnerability
 from vulnerabilities.models import VulnerabilityAnalysisMixin
+from vulnerabilities.models import VulnerabilityAnalysis
+
 
 RELATION_LICENSE_EXPRESSION_HELP_TEXT = _(
     "The License Expression assigned to a DejaCode Product Package or Product "
@@ -520,10 +522,10 @@ class Product(BaseProductMixin, FieldChangesMixin, KeywordsMixin, DataspacedMode
 
         if prefetch_related_packages:
             package_qs = Package.objects.filter(product=self).only_rendering_fields()
-            analysis_qs = ProductVulnerabilityAnalysis.objects.filter(product=self)
+            analysis_qs = VulnerabilityAnalysis.objects.filter(product=self)
             vulnerability_qs = vulnerability_qs.prefetch_related(
                 models.Prefetch("affected_packages", package_qs),
-                models.Prefetch("product_vulnerability_analyses", analysis_qs),
+                models.Prefetch("vulnerability_analyses", analysis_qs),
             )
 
         return vulnerability_qs
@@ -1472,26 +1474,3 @@ class ProductDependency(HistoryFieldsMixin, DataspacedModel):
                     "The 'for_package' cannot be the same as 'resolved_to_package'."
                 )
         super().save(*args, **kwargs)
-
-
-class ProductVulnerabilityAnalysis(
-    VulnerabilityAnalysisMixin,
-    HistoryUserFieldsMixin,
-    DataspacedModel,
-):
-    product = models.ForeignKey(
-        to="product_portfolio.Product",
-        related_name="vulnerability_analyses",
-        on_delete=models.CASCADE,
-    )
-    vulnerability = models.ForeignKey(
-        to="vulnerabilities.Vulnerability",
-        related_name="product_vulnerability_analyses",
-        on_delete=models.CASCADE,
-    )
-
-    class Meta:
-        unique_together = (("product", "vulnerability"), ("dataspace", "uuid"))
-
-    def __str__(self):
-        return f"{self.vulnerability} analysis in {self.product}."
